+++ Source/Exchange/Gnome_Sort.c
@@ -30,16 +30,16 @@ void Gnome_Sort(long int array[], int length){
 
 /*
 int main(){
-    int array[] = {9,5,3,7,6,4,8,1,2,10},i;
+    long int array[] = {9,5,3,7,6,4,8,1,2,10},i;
     
     for(i = 0; i < 10; i++)
-        printf("%d ",array[i]);
+        printf("%ld ",array[i]);
     printf("\n\n");
     
     Gnome_Sort(array,10);
     
     for(i = 0; i < 10; i++)
-        printf("%d ",array[i]);
+        printf("%ld ",array[i]);
     return 0;
 }
 */
