+++ Source/Exchange/Optimized_Bubble_Sort.c
@@ -33,17 +33,17 @@ void BubbleSortOptmized(long int array[], int length){
 
 /*
 int main(){
-    int array[] = {66,600,410,815,779,233,998,510,734,65},i;
+    long int array[] = {66,600,410,815,779,233,998,510,734,65},i;
     
     for(i = 0; i < 10; i++)
-        printf("%d ",array[i]);
+        printf("%ld ",array[i]);
         
     printf("\n\n");
     
     BubbleSortOptmized(array,10-1);
     
     for(i = 0; i < 10; i++)
-        printf("%d ",array[i]);
+        printf("%ld ",array[i]);
 
     return 0;
 }
