+++ Source/Insertion/Binary_Insertion_Sort.c
@@ -37,17 +37,17 @@ void InsertionSort(long int array[], int length){
 
 /*
 int main(){
-    int array[] = {693,378,660,247,77,223,132,685,873,862},i;
+    long int array[] = {693,378,660,247,77,223,132,685,873,862},i;
     
     for(i = 0; i < 10; i++)
-        printf("%d ",array[i]);
+        printf("%ld ",array[i]);
     
     printf("\n\n");
     
     InsertionSort(array,10);
     
     for(i = 0; i < 10; i++)
-        printf("%d ",array[i]);
+        printf("%ld ",array[i]);
     
     return 0;
 }
