+++ Source/Networks&Concurrent/Bitonic_Sort.c
@@ -34,17 +34,17 @@ void BitonicSort(long int array[], int low, int length, int direction){ // Direc
 
 /*
 int main(){
-	int array[] = {514,166,8,269,767,237,318,980};
+	long int array[] = {514,166,8,269,767,237,318,980};
     
     for(int i = 0; i < 8; i++)
-        printf("%d ",array[i]);
+        printf("%ld ",array[i]);
     
     printf("\n\n");
     
     BitonicSort(array,0,8,1);
     
     for(int i = 0; i < 8; i++)
-        printf("%d ",array[i]);
+        printf("%ld ",array[i]);
 
 	return 0;
 }
