+++ Source/Selection/Double_Selection_Sort.c
@@ -47,16 +47,16 @@ void Double_Selection_Sort(long int array[], int length){
 
 /*
 int main(){
-    int array[] = {5,98,-1,0,11,321,654,555,44,2};
+    long int array[] = {5,98,-1,0,11,321,654,555,44,2};
     
     for(int i = 0; i < 10; i++)
-        printf("%d ",array[i]);
+        printf("%ld ",array[i]);
     
     printf("\n\n");
     Double_Selection_Sort(array,10);
     
     for(int i = 0; i < 10; i++)
-        printf("%d ",array[i]);
+        printf("%ld ",array[i]);
     
     return 0;
 }
